bnb_backend/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.10
dj-rest-auth==4.0.1
django-allauth==0.53.0
django-cors-headers==4.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.4.0
pillow==10.3.0
channels==4.1.0
daphne==4.1.0
